backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.38
alembic==1.15.1
pydantic[email]==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
asyncpg==0.30.0
redis==5.2.1
python-dotenv==1.0.1
geoalchemy2==0.17.1
httpx==0.28.1
slowapi==0.1.9
loguru==0.7.3
python-slugify==8.0.4
websockets==14.2
bcrypt==4.2.1
cryptography==44.0.1
aiosqlite==0.21.0
